bolsas_render_deploy/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-Login==0.6.2
Flask-Admin==1.6.1
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
APScheduler==3.10.1
yfinance==0.2.24
pandas==2.2.2
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
Flask-SocketIO==5.3.2
python-socketio==5.9.0
eventlet==0.33.3
